+++ packages/linked_scroll_controller/lib/linked_scroll_controller.dart
@@ -88,6 +88,20 @@ class LinkedScrollControllerGroup {
   void resetScroll() {
     jumpTo(0.0);
   }
+
+  /// Applies the viewportDimension of all linked controllers to [value].
+  void applyViewportDimension(double value) {
+    for (final controller in _attachedControllers) {
+      controller.position.applyViewportDimension(value);
+    }
+  }
+
+  /// Notifies the scroll listener of all linked controllers.
+  void notifyListeners() {
+    for (final controller in _attachedControllers) {
+      controller.position.notifyListeners();
+    }
+  }
 }
 
 /// This class provides change notification for [LinkedScrollControllerGroup]'s