+++ packages/linked_scroll_controller/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.1
+
+* Add `LinkedScrollControllerGroup.applyViewportDimension` method that applies  
+  the viewportDimension of all linked controllers.
+* Add `LinkedScrollControllerGroup.notifyListeners` method that notifies  
+  the listener of all linked controllers.
+
 # 0.2.0
 
   * Stable release for null safety.